learner-server/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.13.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py38h294d835_0
amqp=5.0.6=pypi_0
argon2-cffi=20.1.0=py38h294d835_2
async-timeout=3.0.1=py_1000
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
billiard=3.6.4.0=pypi_0
blas=1.0=mkl
bleach=3.3.1=pyhd8ed1ab_0
blinker=1.4=py_1
brotlipy=0.7.0=py38h294d835_1001
ca-certificates=2021.7.5=haa95532_1
cachetools=4.2.2=pyhd8ed1ab_0
celery=5.1.1=pypi_0
certifi=2021.5.30=py38haa95532_0
cffi=1.14.6=py38hd8c33c5_0
chardet=4.0.0=py38haa244fe_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=7.1.2=pypi_0
click-didyoumean=0.0.3=pypi_0
click-plugins=1.1.1=pypi_0
click-repl=0.2.0=pypi_0
cloudpickle=1.6.0=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=3.4.7=py38hd7da0ea_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=pypi_0
dataclasses=0.8=pyhc8e2a94_1
debugpy=1.4.1=py38h885f38d_0
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
dm-tree=0.1.6=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
future=0.18.2=pypi_0
google-auth=1.34.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.1=py_2
greenlet=1.1.1=pypi_0
grpcio=1.38.1=py38he5377a8_0
gym=0.18.3=pypi_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.6.3=py38haa244fe_0
intel-openmp=2021.3.0=haa95532_3372
ipykernel=6.0.3=py38h43734a8_0
ipython=7.25.0=py38h43734a8_1
ipython_genutils=0.2.0=py_1
jedi=0.18.0=py38haa244fe_2
jinja2=3.0.1=pyhd8ed1ab_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py38haa244fe_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
kiwisolver=1.3.1=pypi_0
kombu=5.1.0=pypi_0
libprotobuf=3.17.2=h7755175_1
libsodium=1.0.18=h8d14728_1
libuv=1.40.0=he774522_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=2.0.1=py38h294d835_0
matplotlib=3.4.2=pypi_0
matplotlib-inline=0.1.2=pyhd8ed1ab_2
mistune=0.8.4=py38h294d835_1004
mkl=2021.3.0=haa95532_524
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.0=py38h277e83a_2
mkl_random=1.2.2=py38hf11a4ad_0
msys2-conda-epoch=20160418=1
multidict=5.1.0=py38h294d835_1
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.1.0=py38haa244fe_0
nbformat=5.1.3=pyhd8ed1ab_0
nest-asyncio=1.5.1=pyhd8ed1ab_0
ninja=1.10.2=h6d14046_1
notebook=6.4.0=pyha770c72_0
numpy=1.17.4=pypi_0
numpy-base=1.20.3=py38hc2deb75_0
oauthlib=3.1.1=pyhd8ed1ab_0
opencv-python=4.5.3.56=pypi_0
openssl=1.1.1k=h2bbff1b_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.1=pypi_0
pandoc=2.14.1=h8ffe710_0
pandocfilters=1.4.2=py_1
parso=0.8.2=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=8.2.0=pypi_0
pip=21.1.3=py38haa95532_0
prometheus_client=0.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.19=pyha770c72_0
protobuf=3.17.2=py38h885f38d_0
psycopg2=2.9.1=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.19=pypi_0
pygame=2.0.1=pypi_0
pyglet=1.5.15=pypi_0
pygments=2.9.0=pyhd8ed1ab_0
pyjwt=2.1.0=pyhd8ed1ab_0
pymunk=5.6.0=pypi_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py38h294d835_2
pysocks=1.7.1=py38haa244fe_3
python=3.8.10=hdbf39b2_7
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.19.0=pypi_0
python_abi=3.8=2_cp38
pytorch=1.8.1=py3.8_cuda10.1_cudnn7_0
pytz=2021.1=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywin32=300=py38h294d835_0
pywinpty=0.5.7=py38_0
pyzmq=22.1.0=py38h09162b1_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
scipy=1.7.0=pypi_0
send2trash=1.7.1=pyhd8ed1ab_0
setuptools=52.0.0=py38haa95532_0
six=1.13.0=pypi_0
sqlalchemy=1.4.22=pypi_0
sqlite=3.35.4=h2bbff1b_0
stable-baselines3=1.1.0a10=dev_0
tactic-game-gym=0.0.1=dev_0
tensorboard=2.5.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py38haa244fe_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tensorflow-model-optimization=0.6.0=pypi_0
terminado=0.9.4=py38haa95532_0
testpath=0.5.0=pyhd8ed1ab_0
tornado=6.1=py38h294d835_1
tqdm=4.62.0=pypi_0
traitlets=5.0.5=py_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vine=5.0.0=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38_0
winpty=0.4.3=4
yarl=1.6.3=py38h294d835_2
zeromq=4.3.4=h0e60522_0
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h62dcd97_1010
